requirements.txt
-i https://pypi.org/simple
anyio==4.6.2.post1; python_version >= '3.9'
attrs==24.2.0; python_version >= '3.7'
boltons==20.2.1
cattrs==24.1.2; python_version >= '3.8'
certifi==2024.8.30; python_version >= '3.6'
datargs==1.1.0; python_version >= '3.7' and python_version < '4.0'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.27.2; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
linkify-it-py==2.0.3; python_version >= '3.7'
markdown-it-py[linkify,plugins]==3.0.0; python_version >= '3.8'
mdit-py-plugins==0.4.2; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
montydb==2.5.3; python_version >= '3.7'
platformdirs==4.3.6; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
rich==13.9.4; python_full_version >= '3.8.0'
sniffio==1.3.1; python_version >= '3.7'
textual==0.87.1; python_full_version >= '3.8.1' and python_full_version < '4.0.0'
typing-extensions==4.12.2; python_version >= '3.8'
uc-micro-py==1.0.3; python_version >= '3.7'
